Refactor : 개별 주류 상세 내용 패키지 분리
1. 카테고리별 주류 상세 내용 분리
2. 바 그래프 그리는 로직 분리
