+++ app/src/main/java/com/goody/dalda/ui/liquor_details/component/DetailLiquorInfoSection.kt
@@ -1,27 +1,21 @@
 package com.goody.dalda.ui.liquor_details.component
 
-import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.remember
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.res.colorResource
-import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.goody.dalda.R
 import com.goody.dalda.data.AlcoholData
+import com.goody.dalda.ui.liquor_details.component.detail_info_component.DetailSectionBeer
+import com.goody.dalda.ui.liquor_details.component.detail_info_component.DetailSectionSake
+import com.goody.dalda.ui.liquor_details.component.detail_info_component.DetailSectionSoju
+import com.goody.dalda.ui.liquor_details.component.detail_info_component.DetailSectionTraditionalLiquor
+import com.goody.dalda.ui.liquor_details.component.detail_info_component.DetailSectionWhiskey
+import com.goody.dalda.ui.liquor_details.component.detail_info_component.DetailSectionWine
 import com.goody.dalda.ui.theme.DaldaTextStyle
-import java.text.DecimalFormat
-
-private const val BAR_COUNT = 5
 
 @Composable
 fun DetailLiquorInfoSection(
@@ -73,359 +67,6 @@ fun DetailLiquorInfoSection(
     }
 }
 
-@Composable
-fun DetailSectionBeer(
-    alcoholData: AlcoholData.Beer,
-    modifier: Modifier = Modifier,
-) {
-    val valueList =
-        listOf(
-            "아로마" to alcoholData.aroma,
-            "색감" to alcoholData.appearance,
-            "맛" to alcoholData.flavor,
-            "바디감" to alcoholData.mouthfeel,
-        ).map { (label, value) -> label to value / 2 }
-
-    val infoList =
-        listOf(
-            "스타일" to alcoholData.type,
-            "제조지역" to alcoholData.country,
-        )
-
-    Column(
-        modifier = modifier,
-        verticalArrangement = Arrangement.spacedBy(12.dp),
-    ) {
-        Text(
-            modifier = Modifier.fillMaxWidth(),
-            text = "술정보",
-            style = DaldaTextStyle.h2,
-        )
-
-        for (info in infoList) {
-            TextTitleValue(
-                modifier = Modifier,
-                title = info.first,
-                value = info.second,
-            )
-        }
-
-        Text(
-            modifier = Modifier.fillMaxWidth(),
-            text = "맛표현",
-            style = DaldaTextStyle.h2,
-        )
-
-        for (value in valueList) {
-            DrawBarGraphWithTitle(
-                modifier = Modifier,
-                value = value,
-            )
-        }
-    }
-}
-
-@Composable
-fun DetailSectionWine(
-    alcoholData: AlcoholData.Wine,
-    modifier: Modifier = Modifier,
-) {
-    val graphDataList =
-        listOf(
-            "단맛" to alcoholData.sugar,
-            "신맛" to alcoholData.acid,
-            "바디감" to alcoholData.mouthfeel,
-        )
-
-    val infoList =
-        listOf(
-            "포도종" to alcoholData.ingredient,
-            "종류" to alcoholData.type,
-            "제조지역" to alcoholData.country,
-            "양조장" to alcoholData.winery,
-        )
-
-    Column(
-        modifier = modifier,
-        verticalArrangement = Arrangement.spacedBy(12.dp),
-    ) {
-        Text(
-            modifier = Modifier.fillMaxWidth(),
-            text = "술정보",
-            style = DaldaTextStyle.h2,
-        )
-
-        Text(
-            modifier =
-            Modifier
-                .background(
-                    colorResource(id = R.color.gray_80),
-                    shape = RoundedCornerShape(12.dp),
-                )
-                .padding(12.dp),
-            text = alcoholData.comment,
-            style = DaldaTextStyle.body3,
-        )
-
-        for (info in infoList) {
-            TextTitleValue(
-                modifier = Modifier,
-                title = info.first,
-                value = info.second,
-            )
-        }
-
-        Text(
-            modifier = Modifier.fillMaxWidth(),
-            text = "맛표현",
-            style = DaldaTextStyle.h2,
-        )
-
-        for (value in graphDataList) {
-            DrawBarGraphWithTitle(
-                modifier = Modifier,
-                value = value,
-            )
-        }
-
-        Text(
-            modifier = Modifier.fillMaxWidth(),
-            text = "푸드 페어링",
-            style = DaldaTextStyle.h2,
-        )
-
-        Text(
-            modifier =
-            Modifier
-                .background(
-                    colorResource(id = R.color.primary_pale),
-                    shape = RoundedCornerShape(12.dp),
-                )
-                .padding(12.dp)
-                .fillMaxWidth(),
-            text = alcoholData.pairingFood,
-            style = DaldaTextStyle.body3,
-        )
-    }
-}
-
-@Composable
-fun DetailSectionTraditionalLiquor(
-    alcoholData: AlcoholData.TraditionalLiquor,
-    modifier: Modifier = Modifier,
-) {
-    val infoList =
-        listOf(
-            "스타일" to alcoholData.type,
-            "재료" to alcoholData.ingredient,
-            "양조장" to alcoholData.brewery,
-        )
-
-    Column(
-        modifier = modifier,
-        verticalArrangement = Arrangement.spacedBy(12.dp),
-    ) {
-        Text(
-            modifier = Modifier.fillMaxWidth(),
-            text = "술정보",
-            style = DaldaTextStyle.h2,
-        )
-
-        Text(
-            modifier =
-            Modifier
-                .background(
-                    colorResource(id = R.color.gray_80),
-                    shape = RoundedCornerShape(12.dp),
-                )
-                .padding(12.dp),
-            text = alcoholData.comment,
-            style = DaldaTextStyle.body3,
-        )
-
-        for (info in infoList) {
-            TextTitleValue(
-                modifier = Modifier,
-                title = info.first,
-                value = info.second,
-            )
-        }
-
-        Text(
-            modifier = Modifier.fillMaxWidth(),
-            text = "푸드 페어링",
-            style = DaldaTextStyle.h2,
-        )
-
-        Text(
-            modifier =
-            Modifier
-                .background(
-                    colorResource(id = R.color.primary_pale),
-                    shape = RoundedCornerShape(12.dp),
-                )
-                .padding(12.dp)
-                .fillMaxWidth(),
-            text = alcoholData.pairingFood,
-            style = DaldaTextStyle.body3,
-        )
-    }
-}
-
-@Composable
-fun DetailSectionWhiskey(
-    alcoholData: AlcoholData.Whisky,
-    modifier: Modifier = Modifier,
-) {
-    val dec = DecimalFormat("#,###")
-
-    val infoList =
-        listOf(
-            "권장 소비자가" to dec.format(alcoholData.price) + "원",
-            "종류" to alcoholData.type,
-            "제조지역" to alcoholData.country,
-        )
-
-    val tasteInfoList =
-        listOf(
-            "아로마" to alcoholData.aroma,
-            "맛" to alcoholData.taste,
-            "마무리" to alcoholData.finish,
-        )
-
-    Column(
-        modifier = modifier,
-        verticalArrangement = Arrangement.spacedBy(12.dp),
-    ) {
-        Text(
-            modifier = Modifier.fillMaxWidth(),
-            text = "술정보",
-            style = DaldaTextStyle.h2,
-        )
-
-        for (info in infoList) {
-            TextTitleValue(
-                modifier = Modifier,
-                title = info.first,
-                value = info.second,
-            )
-        }
-
-        Text(
-            modifier = Modifier.fillMaxWidth(),
-            text = "맛 표현",
-            style = DaldaTextStyle.h2,
-        )
-
-        for (tasteInfo in tasteInfoList) {
-            TextTitleValue(
-                modifier = Modifier,
-                title = tasteInfo.first,
-                value = tasteInfo.second,
-            )
-        }
-    }
-}
-
-@Composable
-fun DetailSectionSake(
-    alcoholData: AlcoholData.Sake,
-    modifier: Modifier = Modifier,
-) {
-    val dec = DecimalFormat("#,###")
-
-    val infoList =
-        listOf(
-            "권장 소비자가" to dec.format(alcoholData.price) + "원",
-            "제조지역" to alcoholData.country,
-        )
-
-    val tasteInfoList =
-        listOf(
-            "아로마" to alcoholData.aroma,
-            "맛" to alcoholData.taste,
-            "마무리" to alcoholData.finish,
-        )
-
-    Column(
-        modifier = modifier,
-        verticalArrangement = Arrangement.spacedBy(12.dp),
-    ) {
-        Text(
-            modifier = Modifier.fillMaxWidth(),
-            text = "술정보",
-            style = DaldaTextStyle.h2,
-        )
-
-        for (info in infoList) {
-            TextTitleValue(
-                modifier = Modifier,
-                title = info.first,
-                value = info.second,
-            )
-        }
-
-        Text(
-            modifier = Modifier.fillMaxWidth(),
-            text = "맛 표현",
-            style = DaldaTextStyle.h2,
-        )
-
-        for (tasteInfo in tasteInfoList) {
-            TextTitleValue(
-                modifier = Modifier,
-                title = tasteInfo.first,
-                value = tasteInfo.second,
-            )
-        }
-    }
-}
-
-@Composable
-fun DetailSectionSoju(
-    alcoholData: AlcoholData.Soju,
-    modifier: Modifier = Modifier,
-) {
-    val dec = DecimalFormat("#,###")
-
-    val infoList =
-        listOf(
-            "권장 소비자가" to dec.format(alcoholData.price) + "원",
-            "제조지역" to alcoholData.country,
-        )
-
-    Column(
-        modifier = modifier,
-        verticalArrangement = Arrangement.spacedBy(12.dp),
-    ) {
-        Text(
-            modifier = Modifier.fillMaxWidth(),
-            text = "술정보",
-            style = DaldaTextStyle.h2,
-        )
-
-        Text(
-            modifier =
-            Modifier
-                .background(
-                    colorResource(id = R.color.gray_80),
-                    shape = RoundedCornerShape(12.dp),
-                )
-                .padding(12.dp)
-                .fillMaxWidth(),
-            text = alcoholData.comment,
-            style = DaldaTextStyle.body3,
-        )
-
-        for (info in infoList) {
-            TextTitleValue(
-                modifier = Modifier,
-                title = info.first,
-                value = info.second,
-            )
-        }
-    }
-}
 
 @Composable
 fun TextTitleValue(
@@ -446,73 +87,6 @@ fun TextTitleValue(
             modifier = Modifier.weight(3f),
             text = value,
             style = DaldaTextStyle.body3,
-        )
-    }
-}
-
-@Composable
-fun DrawBarGraph(
-    value: Float,
-    barCount: Int,
-    modifier: Modifier = Modifier,
-) {
-    val fillWeightList = remember(value, barCount) {
-        List(barCount) { index ->
-            minOf(maxOf(value - index, 0f), 1f)
-        }
-    }
-
-    var barGraphState: BarGraphState
-
-    Row(
-        modifier = modifier,
-    ) {
-        for (i in 0 until barCount) {
-            barGraphState = when (i) {
-                0 -> {
-                    BarGraphState.Start
-                }
-
-                in 1 until barCount - 1 -> {
-                    BarGraphState.Middle
-                }
-
-                else -> {
-                    BarGraphState.End
-                }
-            }
-            BarGraph(
-                value = fillWeightList[i],
-                state = barGraphState,
-                height = 20,
-                color = colorResource(id = R.color.primary),
-                modifier = Modifier.weight(1f),
-            )
-        }
-    }
-}
-
-@Composable
-fun DrawBarGraphWithTitle(
-    value: Pair<String, Float>,
-    modifier: Modifier = Modifier,
-) {
-    Row(
-        modifier = modifier,
-        horizontalArrangement = Arrangement.spacedBy(12.dp),
-    ) {
-        Text(
-            modifier = Modifier
-                .weight(1f)
-                .align(Alignment.CenterVertically),
-            textAlign = TextAlign.End,
-            text = value.first,
-            style = DaldaTextStyle.body2,
-        )
-        DrawBarGraph(
-            modifier = Modifier.weight(7f),
-            value = value.second,
-            barCount = BAR_COUNT,
         )
     }
 }
